etl/requirements.txt
pandas
psycopg2-binary
pymongo
requests
